Split utils into multiple files
