+++ api/src/catalog/catalog.controllers.js
@@ -1,6 +1,6 @@
 import { FERRY_SECRET } from '../config/constants';
 
-import { fetchCatalog } from '../utils.js';
+import { fetchCatalog } from './catalog.utils';
 
 /**
  * Middleware to verify request headers
