+++ api/src/challenge/challenge.controllers.js
@@ -10,7 +10,11 @@ import {
 
 import { requestEvalsQuery, verifyEvalsQuery } from './challenge.queries.js';
 
-import { encrypt, decrypt, getRandomInt } from '../utils.js';
+import {
+  encrypt,
+  decrypt,
+  getRandomInt,
+} from '../../challenge/challenge.utils.js/utils.js';
 
 import Student from '../models/student.models.js';
